chore: rename tools package
