+++ internal/helper/set_log_level.go
@@ -1,4 +1,4 @@
-package tools
+package helper
 
 import (
 	"github.com/rs/zerolog"
